+++ app/app.go
@@ -28,6 +28,7 @@ import (
 
 	"github.com/medibloc/panacea-core/x/aol"
 	"github.com/medibloc/panacea-core/x/did"
+	"github.com/medibloc/panacea-core/x/token"
 
 	abci "github.com/tendermint/tendermint/abci/types"
 	cmn "github.com/tendermint/tendermint/libs/common"
@@ -61,6 +62,7 @@ var (
 		supply.AppModuleBasic{},
 		aol.AppModuleBasic{},
 		did.AppModuleBasic{},
+		token.AppModuleBasic{},
 	)
 
 	// module account permissions
@@ -98,6 +100,7 @@ type PanaceaApp struct {
 	paramsKeeper   params.Keeper
 	aolKeeper      aol.Keeper
 	didKeeper      did.Keeper
+	tokenKeeper    token.Keeper
 
 	// the module manager
 	mm *module.Manager
@@ -121,7 +124,7 @@ func NewPanaceaApp(
 	keys := sdk.NewKVStoreKeys(
 		bam.MainStoreKey, auth.StoreKey, staking.StoreKey,
 		supply.StoreKey, mint.StoreKey, distr.StoreKey, slashing.StoreKey,
-		gov.StoreKey, params.StoreKey, aol.StoreKey, did.StoreKey,
+		gov.StoreKey, params.StoreKey, aol.StoreKey, did.StoreKey, token.StoreKey,
 	)
 	tkeys := sdk.NewTransientStoreKeys(staking.TStoreKey, params.TStoreKey)
 
@@ -215,6 +218,11 @@ func NewPanaceaApp(
 		keys[did.StoreKey],
 		app.cdc,
 	)
+	app.tokenKeeper = token.NewKeeper(
+		keys[token.StoreKey],
+		app.cdc,
+		app.bankKeeper,
+	)
 
 	// register the staking hooks
 	// NOTE: stakingKeeper above is passed by reference, so that it will contain these hooks
@@ -238,6 +246,7 @@ func NewPanaceaApp(
 		staking.NewAppModule(app.stakingKeeper, app.distrKeeper, app.accountKeeper, app.supplyKeeper),
 		aol.NewAppModule(app.aolKeeper),
 		did.NewAppModule(app.didKeeper),
+		token.NewAppModule(app.tokenKeeper),
 	)
 
 	// During begin block slashing happens after distr.BeginBlocker so that
@@ -253,7 +262,7 @@ func NewPanaceaApp(
 		genaccounts.ModuleName, distr.ModuleName, staking.ModuleName,
 		auth.ModuleName, bank.ModuleName, slashing.ModuleName, gov.ModuleName,
 		mint.ModuleName, supply.ModuleName, crisis.ModuleName, genutil.ModuleName,
-		aol.ModuleName, did.ModuleName,
+		aol.ModuleName, did.ModuleName, token.ModuleName,
 	)
 
 	app.mm.RegisterInvariants(&app.crisisKeeper)