+++ x/token/types/codec.go
@@ -1,0 +1,18 @@
+package types
+
+import "github.com/cosmos/cosmos-sdk/codec"
+
+// RegisterCodec registers concrete types on Amino codec
+func RegisterCodec(cdc *codec.Codec) {
+	cdc.RegisterConcrete(MsgIssueToken{}, "token/MsgIssueToken", nil)
+}
+
+// ModuleCdc generic sealed codec to be used throughout module
+var ModuleCdc *codec.Codec
+
+func init() {
+	ModuleCdc = codec.New()
+	RegisterCodec(ModuleCdc)
+	codec.RegisterCrypto(ModuleCdc)
+	ModuleCdc.Seal()
+}