+++ x/token/types/querier.go
@@ -1,0 +1,14 @@
+package types
+
+const (
+	QueryToken      = "token"
+	QueryListTokens = "listTokens"
+)
+
+type QueryTokenParams struct {
+	Symbol Symbol
+}
+
+func NewQueryTokenParams(symbol Symbol) *QueryTokenParams {
+	return &QueryTokenParams{Symbol: symbol}
+}